+++ src/utils/clack-utils.ts
@@ -868,6 +868,7 @@ export function isUsingTypeScript() {
 export async function getOrAskForProjectData(
   options: WizardOptions,
   platform?:
+    | 'javascript-angular'
     | 'javascript-nextjs'
     | 'javascript-nuxt'
     | 'javascript-remix'
@@ -1035,6 +1036,7 @@ export async function askForWizardLogin(options: {
   url: string;
   promoCode?: string;
   platform?:
+    | 'javascript-angular'
     | 'javascript-nextjs'
     | 'javascript-nuxt'
     | 'javascript-remix'