+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - feat: Add `coming-from` parameter ([#837](https://github.com/getsentry/sentry-wizard/pull/837))
+- fix: Prevent addition of multiple `sentry:sourcemaps` commands
 
 ## 4.1.0